glow

Module native

Macros

khr_fallback 🔒

Structs

Constants 🔒

Context

DebugCallbackRawPtr 🔒

Store a boxed callback (i.e., Box<Box<dyn FnMut(...)>>) as a raw pointer, so that it can be referenced by the C API and later converted back into a Box and dropped.

NativeBuffer

NativeFence

NativeFramebuffer

NativeProgram

NativeQuery

NativeRenderbuffer

NativeSampler

NativeShader

NativeTexture

NativeTransformFeedback

NativeUniformLocation

NativeVertexArray

Functions

non_zero_gl_name 🔒

raw_debug_message_callback 🔒