Type Alias glow::Shader

pub type Shader = <Context as HasContext>::Shader;

Implementations

impl Context

pub unsafe fn from_loader_function_cstr<F>(loader_function: F) -> Selfwhere F: FnMut(&CStr) -> *const c_void,

pub unsafe fn from_loader_function<F>(loader_function: F) -> Selfwhere F: FnMut(&str) -> *const c_void,

pub unsafe fn create_texture_from_gl_name(gl_name: c_uint) -> NativeTexture

👎Deprecated: Use the NativeTexture constructor instead

Creates a texture from an external GL name.

This can be useful when a texture is created outside of glow (e.g. OpenXR surface) but glow still needs access to it for rendering.

pub unsafe fn create_framebuffer_from_gl_name( gl_name: c_uint ) -> NativeFramebuffer

👎Deprecated: Use the NativeFramebuffer constructor instead

Creates a framebuffer from an external GL name.

This can be useful when a framebuffer is created outside of glow (e.g: via surfman or another crate that supports sharing of buffers between GL contexts), but glow needs to set it as a target.