Crate glow

Structs

ActiveAttribute

ActiveTransformFeedback

ActiveUniform

Context

DebugMessageLogEntry

NativeBuffer

NativeFence

NativeFramebuffer

NativeProgram

NativeQuery

NativeRenderbuffer

NativeSampler

NativeShader

NativeTexture

NativeTransformFeedback

NativeUniformLocation

NativeVertexArray

ProgramBinary

ShaderPrecisionFormat

Version

A version number for a specific component of an OpenGL implementation

Enums

CompressedPixelUnpackData

PixelPackData

PixelUnpackData

Constants

ACTIVE_ATOMIC_COUNTER_BUFFERS

ACTIVE_ATTRIBUTES

ACTIVE_ATTRIBUTE_MAX_LENGTH

ACTIVE_PROGRAM

ACTIVE_RESOURCES

ACTIVE_SUBROUTINES

ACTIVE_SUBROUTINE_MAX_LENGTH

ACTIVE_SUBROUTINE_UNIFORMS

ACTIVE_SUBROUTINE_UNIFORM_LOCATIONS

ACTIVE_SUBROUTINE_UNIFORM_MAX_LENGTH

ACTIVE_TEXTURE

ACTIVE_UNIFORMS

ACTIVE_UNIFORM_BLOCKS

ACTIVE_UNIFORM_BLOCK_MAX_NAME_LENGTH

ACTIVE_UNIFORM_MAX_LENGTH

ACTIVE_VARIABLES

ALIASED_LINE_WIDTH_RANGE

ALIASED_POINT_SIZE_RANGE

ALL_BARRIER_BITS

ALL_SHADER_BITS

ALPHA

ALPHA_BITS

ALREADY_SIGNALED

ALWAYS

AND

AND_INVERTED

AND_REVERSE

ANY_SAMPLES_PASSED

ANY_SAMPLES_PASSED_CONSERVATIVE

ARRAY_BUFFER

ARRAY_BUFFER_BINDING

ARRAY_SIZE

ARRAY_STRIDE

ATOMIC_COUNTER_BARRIER_BIT

ATOMIC_COUNTER_BUFFER

ATOMIC_COUNTER_BUFFER_ACTIVE_ATOMIC_COUNTERS

ATOMIC_COUNTER_BUFFER_ACTIVE_ATOMIC_COUNTER_INDICES

ATOMIC_COUNTER_BUFFER_BINDING

ATOMIC_COUNTER_BUFFER_DATA_SIZE

ATOMIC_COUNTER_BUFFER_INDEX

ATOMIC_COUNTER_BUFFER_REFERENCED_BY_COMPUTE_SHADER

ATOMIC_COUNTER_BUFFER_REFERENCED_BY_FRAGMENT_SHADER

ATOMIC_COUNTER_BUFFER_REFERENCED_BY_GEOMETRY_SHADER

ATOMIC_COUNTER_BUFFER_REFERENCED_BY_TESS_CONTROL_SHADER

ATOMIC_COUNTER_BUFFER_REFERENCED_BY_TESS_EVALUATION_SHADER

ATOMIC_COUNTER_BUFFER_REFERENCED_BY_VERTEX_SHADER

ATOMIC_COUNTER_BUFFER_SIZE

ATOMIC_COUNTER_BUFFER_START

ATTACHED_SHADERS

AUTO_GENERATE_MIPMAP

BACK

BACK_LEFT

BACK_RIGHT

BGR

BGRA

BGRA_INTEGER

BGR_INTEGER

BLEND

BLEND_COLOR

BLEND_DST

BLEND_DST_ALPHA

BLEND_DST_RGB

BLEND_EQUATION

BLEND_EQUATION_ALPHA

BLEND_EQUATION_RGB

BLEND_SRC

BLEND_SRC_ALPHA

BLEND_SRC_RGB

BLOCK_INDEX

BLUE

BLUE_BITS

BLUE_INTEGER

BOOL

BOOL_VEC2

BOOL_VEC3

BOOL_VEC4

BUFFER

BUFFER_ACCESS

BUFFER_ACCESS_FLAGS

BUFFER_BINDING

BUFFER_DATA_SIZE

BUFFER_IMMUTABLE_STORAGE

BUFFER_MAPPED

BUFFER_MAP_LENGTH

BUFFER_MAP_OFFSET

BUFFER_MAP_POINTER

BUFFER_SIZE

BUFFER_STORAGE_FLAGS

BUFFER_UPDATE_BARRIER_BIT

BUFFER_USAGE

BUFFER_VARIABLE

BYTE

CAVEAT_SUPPORT

CCW

CLAMP_READ_COLOR

CLAMP_TO_BORDER

CLAMP_TO_EDGE

CLEAR

CLEAR_BUFFER

CLEAR_TEXTURE

CLIENT_MAPPED_BUFFER_BARRIER_BIT

CLIENT_STORAGE_BIT

CLIPPING_INPUT_PRIMITIVES

CLIPPING_OUTPUT_PRIMITIVES

CLIP_DEPTH_MODE

CLIP_DISTANCE0

CLIP_DISTANCE1

CLIP_DISTANCE2

CLIP_DISTANCE3

CLIP_DISTANCE4

CLIP_DISTANCE5

CLIP_DISTANCE6

CLIP_DISTANCE7

CLIP_ORIGIN

COLOR

COLOR_ATTACHMENT0

COLOR_ATTACHMENT1

COLOR_ATTACHMENT2

COLOR_ATTACHMENT3

COLOR_ATTACHMENT4

COLOR_ATTACHMENT5

COLOR_ATTACHMENT6

COLOR_ATTACHMENT7

COLOR_ATTACHMENT8

COLOR_ATTACHMENT9

COLOR_ATTACHMENT10

COLOR_ATTACHMENT11

COLOR_ATTACHMENT12

COLOR_ATTACHMENT13

COLOR_ATTACHMENT14

COLOR_ATTACHMENT15

COLOR_ATTACHMENT16

COLOR_ATTACHMENT17

COLOR_ATTACHMENT18

COLOR_ATTACHMENT19

COLOR_ATTACHMENT20

COLOR_ATTACHMENT21

COLOR_ATTACHMENT22

COLOR_ATTACHMENT23

COLOR_ATTACHMENT24

COLOR_ATTACHMENT25

COLOR_ATTACHMENT26

COLOR_ATTACHMENT27

COLOR_ATTACHMENT28

COLOR_ATTACHMENT29

COLOR_ATTACHMENT30

COLOR_ATTACHMENT31

COLOR_BUFFER_BIT

COLOR_CLEAR_VALUE

COLOR_COMPONENTS

COLOR_ENCODING

COLOR_LOGIC_OP

COLOR_RENDERABLE

COLOR_WRITEMASK

COMMAND_BARRIER_BIT

COMPARE_REF_TO_TEXTURE

COMPATIBLE_SUBROUTINES

COMPILE_STATUS

COMPLETION_STATUS

COMPRESSED_R11_EAC

COMPRESSED_RED

COMPRESSED_RED_RGTC1

COMPRESSED_RG

COMPRESSED_RG11_EAC

COMPRESSED_RGB

COMPRESSED_RGB8_ETC2

COMPRESSED_RGB8_PUNCHTHROUGH_ALPHA1_ETC2

COMPRESSED_RGBA

COMPRESSED_RGBA8_ETC2_EAC

COMPRESSED_RGBA_ASTC_4x4_KHR

COMPRESSED_RGBA_ASTC_5x4_KHR

COMPRESSED_RGBA_ASTC_5x5_KHR

COMPRESSED_RGBA_ASTC_6x5_KHR

COMPRESSED_RGBA_ASTC_6x6_KHR

COMPRESSED_RGBA_ASTC_8x5_KHR

COMPRESSED_RGBA_ASTC_8x6_KHR

COMPRESSED_RGBA_ASTC_8x8_KHR

COMPRESSED_RGBA_ASTC_10x5_KHR

COMPRESSED_RGBA_ASTC_10x6_KHR

COMPRESSED_RGBA_ASTC_10x8_KHR

COMPRESSED_RGBA_ASTC_10x10_KHR

COMPRESSED_RGBA_ASTC_12x10_KHR

COMPRESSED_RGBA_ASTC_12x12_KHR

COMPRESSED_RGBA_BPTC_UNORM

COMPRESSED_RGBA_S3TC_DXT1_EXT

COMPRESSED_RGBA_S3TC_DXT3_EXT

COMPRESSED_RGBA_S3TC_DXT5_EXT

COMPRESSED_RGB_BPTC_SIGNED_FLOAT

COMPRESSED_RGB_BPTC_UNSIGNED_FLOAT

COMPRESSED_RGB_S3TC_DXT1_EXT

COMPRESSED_RG_RGTC2

COMPRESSED_SIGNED_R11_EAC

COMPRESSED_SIGNED_RED_RGTC1

COMPRESSED_SIGNED_RG11_EAC

COMPRESSED_SIGNED_RG_RGTC2

COMPRESSED_SRGB

COMPRESSED_SRGB8_ALPHA8_ASTC_4x4_KHR

COMPRESSED_SRGB8_ALPHA8_ASTC_5x4_KHR

COMPRESSED_SRGB8_ALPHA8_ASTC_5x5_KHR

COMPRESSED_SRGB8_ALPHA8_ASTC_6x5_KHR

COMPRESSED_SRGB8_ALPHA8_ASTC_6x6_KHR

COMPRESSED_SRGB8_ALPHA8_ASTC_8x5_KHR

COMPRESSED_SRGB8_ALPHA8_ASTC_8x6_KHR

COMPRESSED_SRGB8_ALPHA8_ASTC_8x8_KHR

COMPRESSED_SRGB8_ALPHA8_ASTC_10x5_KHR

COMPRESSED_SRGB8_ALPHA8_ASTC_10x6_KHR

COMPRESSED_SRGB8_ALPHA8_ASTC_10x8_KHR

COMPRESSED_SRGB8_ALPHA8_ASTC_10x10_KHR

COMPRESSED_SRGB8_ALPHA8_ASTC_12x10_KHR

COMPRESSED_SRGB8_ALPHA8_ASTC_12x12_KHR

COMPRESSED_SRGB8_ALPHA8_ETC2_EAC

COMPRESSED_SRGB8_ETC2

COMPRESSED_SRGB8_PUNCHTHROUGH_ALPHA1_ETC2

COMPRESSED_SRGB_ALPHA

COMPRESSED_SRGB_ALPHA_BPTC_UNORM

COMPRESSED_SRGB_ALPHA_S3TC_DXT1_EXT

COMPRESSED_SRGB_ALPHA_S3TC_DXT3_EXT

COMPRESSED_SRGB_ALPHA_S3TC_DXT5_EXT

COMPRESSED_SRGB_S3TC_DXT1_EXT

COMPRESSED_TEXTURE_FORMATS

COMPUTE_SHADER

COMPUTE_SHADER_BIT

COMPUTE_SHADER_INVOCATIONS

COMPUTE_SUBROUTINE

COMPUTE_SUBROUTINE_UNIFORM

COMPUTE_TEXTURE

COMPUTE_WORK_GROUP_SIZE

CONDITION_SATISFIED

CONSTANT_ALPHA

CONSTANT_COLOR

CONTEXT_COMPATIBILITY_PROFILE_BIT

CONTEXT_CORE_PROFILE_BIT

CONTEXT_FLAGS

CONTEXT_FLAG_DEBUG_BIT

CONTEXT_FLAG_FORWARD_COMPATIBLE_BIT

CONTEXT_FLAG_NO_ERROR_BIT

CONTEXT_FLAG_ROBUST_ACCESS_BIT

CONTEXT_LOST

CONTEXT_PROFILE_MASK

CONTEXT_RELEASE_BEHAVIOR

CONTEXT_RELEASE_BEHAVIOR_FLUSH

COPY

COPY_INVERTED

COPY_READ_BUFFER

COPY_READ_BUFFER_BINDING

COPY_WRITE_BUFFER

COPY_WRITE_BUFFER_BINDING

CULL_FACE

CULL_FACE_MODE

CURRENT_PROGRAM

CURRENT_QUERY

CURRENT_VERTEX_ATTRIB

CW

DEBUG_CALLBACK_FUNCTION

DEBUG_CALLBACK_USER_PARAM

DEBUG_GROUP_STACK_DEPTH

DEBUG_LOGGED_MESSAGES

DEBUG_NEXT_LOGGED_MESSAGE_LENGTH

DEBUG_OUTPUT

DEBUG_OUTPUT_SYNCHRONOUS

DEBUG_SEVERITY_HIGH

DEBUG_SEVERITY_LOW

DEBUG_SEVERITY_MEDIUM

DEBUG_SEVERITY_NOTIFICATION

DEBUG_SOURCE_API

DEBUG_SOURCE_APPLICATION

DEBUG_SOURCE_OTHER

DEBUG_SOURCE_SHADER_COMPILER

DEBUG_SOURCE_THIRD_PARTY

DEBUG_SOURCE_WINDOW_SYSTEM

DEBUG_TYPE_DEPRECATED_BEHAVIOR

DEBUG_TYPE_ERROR

DEBUG_TYPE_MARKER

DEBUG_TYPE_OTHER

DEBUG_TYPE_PERFORMANCE

DEBUG_TYPE_POP_GROUP

DEBUG_TYPE_PORTABILITY

DEBUG_TYPE_PUSH_GROUP

DEBUG_TYPE_UNDEFINED_BEHAVIOR

DECR

DECR_WRAP

DELETE_STATUS

DEPTH

DEPTH24_STENCIL8

DEPTH32F_STENCIL8

DEPTH_ATTACHMENT

DEPTH_BITS

DEPTH_BUFFER_BIT

DEPTH_CLAMP

DEPTH_CLEAR_VALUE

DEPTH_COMPONENT

DEPTH_COMPONENT16

DEPTH_COMPONENT24

DEPTH_COMPONENT32

DEPTH_COMPONENT32F

DEPTH_COMPONENTS

DEPTH_FUNC

DEPTH_RANGE

DEPTH_RENDERABLE

DEPTH_STENCIL

DEPTH_STENCIL_ATTACHMENT

DEPTH_STENCIL_TEXTURE_MODE

DEPTH_TEST

DEPTH_WRITEMASK

DISPATCH_INDIRECT_BUFFER

DISPATCH_INDIRECT_BUFFER_BINDING

DISPLAY_LIST

DITHER

DONT_CARE

DOUBLE

DOUBLEBUFFER

DOUBLE_MAT2

DOUBLE_MAT3

DOUBLE_MAT4

DOUBLE_MAT2x3

DOUBLE_MAT2x4

DOUBLE_MAT3x2

DOUBLE_MAT3x4

DOUBLE_MAT4x2

DOUBLE_MAT4x3

DOUBLE_VEC2

DOUBLE_VEC3

DOUBLE_VEC4

DRAW_BUFFER

DRAW_BUFFER0

DRAW_BUFFER1

DRAW_BUFFER2

DRAW_BUFFER3

DRAW_BUFFER4

DRAW_BUFFER5

DRAW_BUFFER6

DRAW_BUFFER7

DRAW_BUFFER8

DRAW_BUFFER9

DRAW_BUFFER10

DRAW_BUFFER11

DRAW_BUFFER12

DRAW_BUFFER13

DRAW_BUFFER14

DRAW_BUFFER15

DRAW_FRAMEBUFFER

DRAW_FRAMEBUFFER_BINDING

DRAW_INDIRECT_BUFFER

DRAW_INDIRECT_BUFFER_BINDING

DST_ALPHA

DST_COLOR

DYNAMIC_COPY

DYNAMIC_DRAW

DYNAMIC_READ

DYNAMIC_STORAGE_BIT

ELEMENT_ARRAY_BARRIER_BIT

ELEMENT_ARRAY_BUFFER

ELEMENT_ARRAY_BUFFER_BINDING

EQUAL

EQUIV

EXTENSIONS

FALSE

FASTEST

FILL

FILTER

FIRST_VERTEX_CONVENTION

FIXED

FIXED_ONLY

FLOAT

FLOAT_32_UNSIGNED_INT_24_8_REV

FLOAT_MAT2

FLOAT_MAT3

FLOAT_MAT4

FLOAT_MAT2x3

FLOAT_MAT2x4

FLOAT_MAT3x2

FLOAT_MAT3x4

FLOAT_MAT4x2

FLOAT_MAT4x3

FLOAT_VEC2

FLOAT_VEC3

FLOAT_VEC4

FRACTIONAL_EVEN

FRACTIONAL_ODD

FRAGMENT_INTERPOLATION_OFFSET_BITS

FRAGMENT_SHADER

FRAGMENT_SHADER_BIT

FRAGMENT_SHADER_DERIVATIVE_HINT

FRAGMENT_SHADER_INVOCATIONS

FRAGMENT_SUBROUTINE

FRAGMENT_SUBROUTINE_UNIFORM

FRAGMENT_TEXTURE

FRAMEBUFFER

FRAMEBUFFER_ATTACHMENT_ALPHA_SIZE

FRAMEBUFFER_ATTACHMENT_BLUE_SIZE

FRAMEBUFFER_ATTACHMENT_COLOR_ENCODING

FRAMEBUFFER_ATTACHMENT_COMPONENT_TYPE

FRAMEBUFFER_ATTACHMENT_DEPTH_SIZE

FRAMEBUFFER_ATTACHMENT_GREEN_SIZE

FRAMEBUFFER_ATTACHMENT_LAYERED

FRAMEBUFFER_ATTACHMENT_OBJECT_NAME

FRAMEBUFFER_ATTACHMENT_OBJECT_TYPE

FRAMEBUFFER_ATTACHMENT_RED_SIZE

FRAMEBUFFER_ATTACHMENT_STENCIL_SIZE

FRAMEBUFFER_ATTACHMENT_TEXTURE_CUBE_MAP_FACE

FRAMEBUFFER_ATTACHMENT_TEXTURE_LAYER

FRAMEBUFFER_ATTACHMENT_TEXTURE_LEVEL

FRAMEBUFFER_BARRIER_BIT

FRAMEBUFFER_BINDING

FRAMEBUFFER_BLEND

FRAMEBUFFER_COMPLETE

FRAMEBUFFER_DEFAULT

FRAMEBUFFER_DEFAULT_FIXED_SAMPLE_LOCATIONS

FRAMEBUFFER_DEFAULT_HEIGHT

FRAMEBUFFER_DEFAULT_LAYERS

FRAMEBUFFER_DEFAULT_SAMPLES

FRAMEBUFFER_DEFAULT_WIDTH

FRAMEBUFFER_INCOMPLETE_ATTACHMENT

FRAMEBUFFER_INCOMPLETE_DIMENSIONS

FRAMEBUFFER_INCOMPLETE_DRAW_BUFFER

FRAMEBUFFER_INCOMPLETE_LAYER_TARGETS

FRAMEBUFFER_INCOMPLETE_MISSING_ATTACHMENT

FRAMEBUFFER_INCOMPLETE_MULTISAMPLE

FRAMEBUFFER_INCOMPLETE_READ_BUFFER

FRAMEBUFFER_RENDERABLE

FRAMEBUFFER_RENDERABLE_LAYERED

FRAMEBUFFER_SRGB

FRAMEBUFFER_UNDEFINED

FRAMEBUFFER_UNSUPPORTED

FRONT

FRONT_AND_BACK

FRONT_FACE

FRONT_LEFT

FRONT_RIGHT

FULL_SUPPORT

FUNC_ADD

FUNC_REVERSE_SUBTRACT

FUNC_SUBTRACT

GENERATE_MIPMAP_HINT

GEOMETRY_INPUT_TYPE

GEOMETRY_OUTPUT_TYPE

GEOMETRY_SHADER

GEOMETRY_SHADER_BIT

GEOMETRY_SHADER_INVOCATIONS

GEOMETRY_SHADER_PRIMITIVES_EMITTED

GEOMETRY_SUBROUTINE

GEOMETRY_SUBROUTINE_UNIFORM

GEOMETRY_TEXTURE

GEOMETRY_VERTICES_OUT

GEQUAL

GET_TEXTURE_IMAGE_FORMAT

GET_TEXTURE_IMAGE_TYPE

GREATER

GREEN

GREEN_BITS

GREEN_INTEGER

GUILTY_CONTEXT_RESET

HALF_FLOAT

HALF_FLOAT_OES

HIGH_FLOAT

HIGH_INT

IMAGE_1D

IMAGE_1D_ARRAY

IMAGE_2D

IMAGE_2D_ARRAY

IMAGE_2D_MULTISAMPLE

IMAGE_2D_MULTISAMPLE_ARRAY

IMAGE_2D_RECT

IMAGE_3D

IMAGE_BINDING_ACCESS

IMAGE_BINDING_FORMAT

IMAGE_BINDING_LAYER

IMAGE_BINDING_LAYERED

IMAGE_BINDING_LEVEL

IMAGE_BINDING_NAME

IMAGE_BUFFER

IMAGE_CLASS_1_X_8

IMAGE_CLASS_1_X_16

IMAGE_CLASS_1_X_32

IMAGE_CLASS_2_X_8

IMAGE_CLASS_2_X_16

IMAGE_CLASS_2_X_32

IMAGE_CLASS_4_X_8

IMAGE_CLASS_4_X_16

IMAGE_CLASS_4_X_32

IMAGE_CLASS_10_10_10_2

IMAGE_CLASS_11_11_10

IMAGE_COMPATIBILITY_CLASS

IMAGE_CUBE

IMAGE_CUBE_MAP_ARRAY

IMAGE_FORMAT_COMPATIBILITY_BY_CLASS

IMAGE_FORMAT_COMPATIBILITY_BY_SIZE

IMAGE_FORMAT_COMPATIBILITY_TYPE

IMAGE_PIXEL_FORMAT

IMAGE_PIXEL_TYPE

IMAGE_TEXEL_SIZE

IMPLEMENTATION_COLOR_READ_FORMAT

IMPLEMENTATION_COLOR_READ_TYPE

INCR

INCR_WRAP

INDEX

INFO_LOG_LENGTH

INNOCENT_CONTEXT_RESET

INT

INTERLEAVED_ATTRIBS

INTERNALFORMAT_ALPHA_SIZE

INTERNALFORMAT_ALPHA_TYPE

INTERNALFORMAT_BLUE_SIZE

INTERNALFORMAT_BLUE_TYPE

INTERNALFORMAT_DEPTH_SIZE

INTERNALFORMAT_DEPTH_TYPE

INTERNALFORMAT_GREEN_SIZE

INTERNALFORMAT_GREEN_TYPE

INTERNALFORMAT_PREFERRED

INTERNALFORMAT_RED_SIZE

INTERNALFORMAT_RED_TYPE

INTERNALFORMAT_SHARED_SIZE

INTERNALFORMAT_STENCIL_SIZE

INTERNALFORMAT_STENCIL_TYPE

INTERNALFORMAT_SUPPORTED

INT_2_10_10_10_REV

INT_IMAGE_1D

INT_IMAGE_1D_ARRAY

INT_IMAGE_2D

INT_IMAGE_2D_ARRAY

INT_IMAGE_2D_MULTISAMPLE

INT_IMAGE_2D_MULTISAMPLE_ARRAY

INT_IMAGE_2D_RECT

INT_IMAGE_3D

INT_IMAGE_BUFFER

INT_IMAGE_CUBE

INT_IMAGE_CUBE_MAP_ARRAY

INT_SAMPLER_1D

INT_SAMPLER_1D_ARRAY

INT_SAMPLER_2D

INT_SAMPLER_2D_ARRAY

INT_SAMPLER_2D_MULTISAMPLE

INT_SAMPLER_2D_MULTISAMPLE_ARRAY

INT_SAMPLER_2D_RECT

INT_SAMPLER_3D

INT_SAMPLER_BUFFER

INT_SAMPLER_CUBE

INT_SAMPLER_CUBE_MAP_ARRAY

INT_VEC2

INT_VEC3

INT_VEC4

INVALID_ENUM

INVALID_FRAMEBUFFER_OPERATION

INVALID_INDEX

INVALID_OPERATION

INVALID_VALUE

INVERT

ISOLINES

IS_PER_PATCH

IS_ROW_MAJOR

KEEP

LAST_VERTEX_CONVENTION

LAYER_PROVOKING_VERTEX

LEFT

LEQUAL

LESS

LINE

LINEAR

LINEAR_MIPMAP_LINEAR

LINEAR_MIPMAP_NEAREST

LINES

LINES_ADJACENCY

LINE_LOOP

LINE_SMOOTH

LINE_SMOOTH_HINT

LINE_STRIP

LINE_STRIP_ADJACENCY

LINE_WIDTH

LINE_WIDTH_GRANULARITY

LINE_WIDTH_RANGE

LINK_STATUS

LOCATION

LOCATION_COMPONENT

LOCATION_INDEX

LOGIC_OP_MODE

LOSE_CONTEXT_ON_RESET

LOWER_LEFT

LOW_FLOAT

LOW_INT

LUMINANCE

LUMINANCE_ALPHA

MAJOR_VERSION

MANUAL_GENERATE_MIPMAP

MAP_COHERENT_BIT

MAP_FLUSH_EXPLICIT_BIT

MAP_INVALIDATE_BUFFER_BIT

MAP_INVALIDATE_RANGE_BIT

MAP_PERSISTENT_BIT

MAP_READ_BIT

MAP_UNSYNCHRONIZED_BIT

MAP_WRITE_BIT

MATRIX_STRIDE

MAX

MAX_3D_TEXTURE_SIZE

MAX_ARRAY_TEXTURE_LAYERS

MAX_ATOMIC_COUNTER_BUFFER_BINDINGS

MAX_ATOMIC_COUNTER_BUFFER_SIZE

MAX_CLIP_DISTANCES

MAX_COLOR_ATTACHMENTS

MAX_COLOR_TEXTURE_SAMPLES

MAX_COMBINED_ATOMIC_COUNTERS

MAX_COMBINED_ATOMIC_COUNTER_BUFFERS

MAX_COMBINED_CLIP_AND_CULL_DISTANCES

MAX_COMBINED_COMPUTE_UNIFORM_COMPONENTS

MAX_COMBINED_DIMENSIONS

MAX_COMBINED_FRAGMENT_UNIFORM_COMPONENTS

MAX_COMBINED_GEOMETRY_UNIFORM_COMPONENTS

MAX_COMBINED_IMAGE_UNIFORMS

MAX_COMBINED_IMAGE_UNITS_AND_FRAGMENT_OUTPUTS

MAX_COMBINED_SHADER_OUTPUT_RESOURCES

MAX_COMBINED_SHADER_STORAGE_BLOCKS

MAX_COMBINED_TESS_CONTROL_UNIFORM_COMPONENTS

MAX_COMBINED_TESS_EVALUATION_UNIFORM_COMPONENTS

MAX_COMBINED_TEXTURE_IMAGE_UNITS

MAX_COMBINED_UNIFORM_BLOCKS

MAX_COMBINED_VERTEX_UNIFORM_COMPONENTS

MAX_COMPUTE_ATOMIC_COUNTERS

MAX_COMPUTE_ATOMIC_COUNTER_BUFFERS

MAX_COMPUTE_IMAGE_UNIFORMS

MAX_COMPUTE_SHADER_STORAGE_BLOCKS

MAX_COMPUTE_SHARED_MEMORY_SIZE

MAX_COMPUTE_TEXTURE_IMAGE_UNITS

MAX_COMPUTE_UNIFORM_BLOCKS

MAX_COMPUTE_UNIFORM_COMPONENTS

MAX_COMPUTE_WORK_GROUP_COUNT

MAX_COMPUTE_WORK_GROUP_INVOCATIONS

MAX_COMPUTE_WORK_GROUP_SIZE

MAX_CUBE_MAP_TEXTURE_SIZE

MAX_CULL_DISTANCES

MAX_DEBUG_GROUP_STACK_DEPTH

MAX_DEBUG_LOGGED_MESSAGES

MAX_DEBUG_MESSAGE_LENGTH

MAX_DEPTH

MAX_DEPTH_TEXTURE_SAMPLES

MAX_DRAW_BUFFERS

MAX_DUAL_SOURCE_DRAW_BUFFERS

MAX_ELEMENTS_INDICES

MAX_ELEMENTS_VERTICES

MAX_ELEMENT_INDEX

MAX_FRAGMENT_ATOMIC_COUNTERS

MAX_FRAGMENT_ATOMIC_COUNTER_BUFFERS

MAX_FRAGMENT_IMAGE_UNIFORMS

MAX_FRAGMENT_INPUT_COMPONENTS

MAX_FRAGMENT_INTERPOLATION_OFFSET

MAX_FRAGMENT_SHADER_STORAGE_BLOCKS

MAX_FRAGMENT_UNIFORM_BLOCKS

MAX_FRAGMENT_UNIFORM_COMPONENTS

MAX_FRAGMENT_UNIFORM_VECTORS

MAX_FRAMEBUFFER_HEIGHT

MAX_FRAMEBUFFER_LAYERS

MAX_FRAMEBUFFER_SAMPLES

MAX_FRAMEBUFFER_WIDTH

MAX_GEOMETRY_ATOMIC_COUNTERS

MAX_GEOMETRY_ATOMIC_COUNTER_BUFFERS

MAX_GEOMETRY_IMAGE_UNIFORMS

MAX_GEOMETRY_INPUT_COMPONENTS

MAX_GEOMETRY_OUTPUT_COMPONENTS

MAX_GEOMETRY_OUTPUT_VERTICES

MAX_GEOMETRY_SHADER_INVOCATIONS

MAX_GEOMETRY_SHADER_STORAGE_BLOCKS

MAX_GEOMETRY_TEXTURE_IMAGE_UNITS

MAX_GEOMETRY_TOTAL_OUTPUT_COMPONENTS

MAX_GEOMETRY_UNIFORM_BLOCKS

MAX_GEOMETRY_UNIFORM_COMPONENTS

MAX_HEIGHT

MAX_IMAGE_SAMPLES

MAX_IMAGE_UNITS

MAX_INTEGER_SAMPLES

MAX_LABEL_LENGTH

MAX_LAYERS

MAX_NAME_LENGTH

MAX_NUM_ACTIVE_VARIABLES

MAX_NUM_COMPATIBLE_SUBROUTINES

MAX_PATCH_VERTICES

MAX_PROGRAM_TEXEL_OFFSET

MAX_PROGRAM_TEXTURE_GATHER_OFFSET

MAX_RECTANGLE_TEXTURE_SIZE

MAX_RENDERBUFFER_SIZE

MAX_SAMPLES

MAX_SAMPLE_MASK_WORDS

MAX_SERVER_WAIT_TIMEOUT

MAX_SHADER_COMPILER_THREADS

MAX_SHADER_STORAGE_BLOCK_SIZE

MAX_SHADER_STORAGE_BUFFER_BINDINGS

MAX_SUBROUTINES

MAX_SUBROUTINE_UNIFORM_LOCATIONS

MAX_TESS_CONTROL_ATOMIC_COUNTERS

MAX_TESS_CONTROL_ATOMIC_COUNTER_BUFFERS

MAX_TESS_CONTROL_IMAGE_UNIFORMS

MAX_TESS_CONTROL_INPUT_COMPONENTS

MAX_TESS_CONTROL_OUTPUT_COMPONENTS

MAX_TESS_CONTROL_SHADER_STORAGE_BLOCKS

MAX_TESS_CONTROL_TEXTURE_IMAGE_UNITS

MAX_TESS_CONTROL_TOTAL_OUTPUT_COMPONENTS

MAX_TESS_CONTROL_UNIFORM_BLOCKS

MAX_TESS_CONTROL_UNIFORM_COMPONENTS

MAX_TESS_EVALUATION_ATOMIC_COUNTERS

MAX_TESS_EVALUATION_ATOMIC_COUNTER_BUFFERS

MAX_TESS_EVALUATION_IMAGE_UNIFORMS

MAX_TESS_EVALUATION_INPUT_COMPONENTS

MAX_TESS_EVALUATION_OUTPUT_COMPONENTS

MAX_TESS_EVALUATION_SHADER_STORAGE_BLOCKS

MAX_TESS_EVALUATION_TEXTURE_IMAGE_UNITS

MAX_TESS_EVALUATION_UNIFORM_BLOCKS

MAX_TESS_EVALUATION_UNIFORM_COMPONENTS

MAX_TESS_GEN_LEVEL

MAX_TESS_PATCH_COMPONENTS

MAX_TEXTURE_BUFFER_SIZE

MAX_TEXTURE_IMAGE_UNITS

MAX_TEXTURE_LOD_BIAS

MAX_TEXTURE_MAX_ANISOTROPY

MAX_TEXTURE_MAX_ANISOTROPY_EXT

MAX_TEXTURE_SIZE

MAX_TRANSFORM_FEEDBACK_BUFFERS

MAX_TRANSFORM_FEEDBACK_INTERLEAVED_COMPONENTS

MAX_TRANSFORM_FEEDBACK_SEPARATE_ATTRIBS

MAX_TRANSFORM_FEEDBACK_SEPARATE_COMPONENTS

MAX_UNIFORM_BLOCK_SIZE

MAX_UNIFORM_BUFFER_BINDINGS

MAX_UNIFORM_LOCATIONS

MAX_VARYING_COMPONENTS

MAX_VARYING_FLOATS

MAX_VARYING_VECTORS

MAX_VERTEX_ATOMIC_COUNTERS

MAX_VERTEX_ATOMIC_COUNTER_BUFFERS

MAX_VERTEX_ATTRIBS

MAX_VERTEX_ATTRIB_BINDINGS

MAX_VERTEX_ATTRIB_RELATIVE_OFFSET

MAX_VERTEX_ATTRIB_STRIDE

MAX_VERTEX_IMAGE_UNIFORMS

MAX_VERTEX_OUTPUT_COMPONENTS

MAX_VERTEX_SHADER_STORAGE_BLOCKS

MAX_VERTEX_STREAMS

MAX_VERTEX_TEXTURE_IMAGE_UNITS

MAX_VERTEX_UNIFORM_BLOCKS

MAX_VERTEX_UNIFORM_COMPONENTS

MAX_VERTEX_UNIFORM_VECTORS

MAX_VIEWPORTS

MAX_VIEWPORT_DIMS

MAX_WIDTH

MEDIUM_FLOAT

MEDIUM_INT

MIN

MINOR_VERSION

MIN_FRAGMENT_INTERPOLATION_OFFSET

MIN_MAP_BUFFER_ALIGNMENT

MIN_PROGRAM_TEXEL_OFFSET

MIN_PROGRAM_TEXTURE_GATHER_OFFSET

MIN_SAMPLE_SHADING_VALUE

MIPMAP

MIRRORED_REPEAT

MIRROR_CLAMP_TO_EDGE

MULTISAMPLE

NAME_LENGTH

NAND

NEAREST

NEAREST_MIPMAP_LINEAR

NEAREST_MIPMAP_NEAREST

NEGATIVE_ONE_TO_ONE

NEVER

NICEST

NONE

NOOP

NOR

NOTEQUAL

NO_ERROR

NO_RESET_NOTIFICATION

NUM_ACTIVE_VARIABLES

NUM_COMPATIBLE_SUBROUTINES

NUM_COMPRESSED_TEXTURE_FORMATS

NUM_EXTENSIONS

NUM_PROGRAM_BINARY_FORMATS

NUM_SAMPLE_COUNTS

NUM_SHADER_BINARY_FORMATS

NUM_SHADING_LANGUAGE_VERSIONS

NUM_SPIR_V_EXTENSIONS

OBJECT_TYPE

OFFSET

ONE

ONE_MINUS_CONSTANT_ALPHA

ONE_MINUS_CONSTANT_COLOR

ONE_MINUS_DST_ALPHA

ONE_MINUS_DST_COLOR

ONE_MINUS_SRC1_ALPHA

ONE_MINUS_SRC1_COLOR

ONE_MINUS_SRC_ALPHA

ONE_MINUS_SRC_COLOR

OR

OR_INVERTED

OR_REVERSE

OUT_OF_MEMORY

PACK_ALIGNMENT

PACK_COMPRESSED_BLOCK_DEPTH

PACK_COMPRESSED_BLOCK_HEIGHT

PACK_COMPRESSED_BLOCK_SIZE

PACK_COMPRESSED_BLOCK_WIDTH

PACK_IMAGE_HEIGHT

PACK_LSB_FIRST

PACK_ROW_LENGTH

PACK_SKIP_IMAGES

PACK_SKIP_PIXELS

PACK_SKIP_ROWS

PACK_SWAP_BYTES

PARAMETER_BUFFER

PARAMETER_BUFFER_BINDING

PATCHES

PATCH_DEFAULT_INNER_LEVEL

PATCH_DEFAULT_OUTER_LEVEL

PATCH_VERTICES

PIXEL_BUFFER_BARRIER_BIT

PIXEL_PACK_BUFFER

PIXEL_PACK_BUFFER_BINDING

PIXEL_UNPACK_BUFFER

PIXEL_UNPACK_BUFFER_BINDING

POINT

POINTS

POINT_FADE_THRESHOLD_SIZE

POINT_SIZE

POINT_SIZE_GRANULARITY

POINT_SIZE_RANGE

POINT_SPRITE_COORD_ORIGIN

POLYGON_MODE

POLYGON_OFFSET_CLAMP

POLYGON_OFFSET_FACTOR

POLYGON_OFFSET_FILL

POLYGON_OFFSET_LINE

POLYGON_OFFSET_POINT

POLYGON_OFFSET_UNITS

POLYGON_SMOOTH

POLYGON_SMOOTH_HINT

PRIMITIVES_GENERATED

PRIMITIVES_SUBMITTED

PRIMITIVE_RESTART

PRIMITIVE_RESTART_FIXED_INDEX

PRIMITIVE_RESTART_FOR_PATCHES_SUPPORTED

PRIMITIVE_RESTART_INDEX

PROGRAM

PROGRAM_BINARY_FORMATS

PROGRAM_BINARY_LENGTH

PROGRAM_BINARY_RETRIEVABLE_HINT

PROGRAM_INPUT

PROGRAM_OUTPUT

PROGRAM_PIPELINE

PROGRAM_PIPELINE_BINDING

PROGRAM_POINT_SIZE

PROGRAM_SEPARABLE

PROVOKING_VERTEX

PROXY_TEXTURE_1D

PROXY_TEXTURE_1D_ARRAY

PROXY_TEXTURE_2D

PROXY_TEXTURE_2D_ARRAY

PROXY_TEXTURE_2D_MULTISAMPLE

PROXY_TEXTURE_2D_MULTISAMPLE_ARRAY

PROXY_TEXTURE_3D

PROXY_TEXTURE_CUBE_MAP

PROXY_TEXTURE_CUBE_MAP_ARRAY

PROXY_TEXTURE_RECTANGLE

QUADS

QUADS_FOLLOW_PROVOKING_VERTEX_CONVENTION

QUERY

QUERY_BUFFER

QUERY_BUFFER_BARRIER_BIT

QUERY_BUFFER_BINDING

QUERY_BY_REGION_NO_WAIT

QUERY_BY_REGION_NO_WAIT_INVERTED

QUERY_BY_REGION_WAIT

QUERY_BY_REGION_WAIT_INVERTED

QUERY_COUNTER_BITS

QUERY_NO_WAIT

QUERY_NO_WAIT_INVERTED

QUERY_RESULT

QUERY_RESULT_AVAILABLE

QUERY_RESULT_NO_WAIT

QUERY_TARGET

QUERY_WAIT

QUERY_WAIT_INVERTED

R8

R3_G3_B2

R8I

R8UI

R8_SNORM

R16

R11F_G11F_B10F

R16F

R16I

R16UI

R16_SNORM

R32F

R32I

R32UI

RASTERIZER_DISCARD

READ_BUFFER

READ_FRAMEBUFFER

READ_FRAMEBUFFER_BINDING

READ_ONLY

READ_PIXELS

READ_PIXELS_FORMAT

READ_PIXELS_TYPE

READ_WRITE

RED

RED_BITS

RED_INTEGER

REFERENCED_BY_COMPUTE_SHADER

REFERENCED_BY_FRAGMENT_SHADER

REFERENCED_BY_GEOMETRY_SHADER

REFERENCED_BY_TESS_CONTROL_SHADER

REFERENCED_BY_TESS_EVALUATION_SHADER

REFERENCED_BY_VERTEX_SHADER

RENDERBUFFER

RENDERBUFFER_ALPHA_SIZE

RENDERBUFFER_BINDING

RENDERBUFFER_BLUE_SIZE

RENDERBUFFER_DEPTH_SIZE

RENDERBUFFER_GREEN_SIZE

RENDERBUFFER_HEIGHT

RENDERBUFFER_INTERNAL_FORMAT

RENDERBUFFER_RED_SIZE

RENDERBUFFER_SAMPLES

RENDERBUFFER_STENCIL_SIZE

RENDERBUFFER_WIDTH

RENDERER

REPEAT

REPLACE

RESET_NOTIFICATION_STRATEGY

RG

RG8

RG8I

RG8UI

RG8_SNORM

RG16

RG16F

RG16I

RG16UI

RG16_SNORM

RG32F

RG32I

RG32UI

RGB

RGB4

RGB5

RGB8

RGB5_A1

RGB8I

RGB8UI

RGB8_SNORM

RGB9_E5

RGB10

RGB12

RGB16

RGB10_A2

RGB10_A2UI

RGB16F

RGB16I

RGB16UI

RGB16_SNORM

RGB32F

RGB32I

RGB32UI

RGB565

RGBA

RGBA2

RGBA4

RGBA8

RGBA8I

RGBA8UI

RGBA8_SNORM

RGBA12

RGBA16

RGBA16F

RGBA16I

RGBA16UI

RGBA16_SNORM

RGBA32F

RGBA32I

RGBA32UI

RGBA_INTEGER

RGB_INTEGER

RG_INTEGER

RIGHT

SAMPLER

SAMPLER_1D

SAMPLER_1D_ARRAY

SAMPLER_1D_ARRAY_SHADOW

SAMPLER_1D_SHADOW

SAMPLER_2D

SAMPLER_2D_ARRAY

SAMPLER_2D_ARRAY_SHADOW

SAMPLER_2D_MULTISAMPLE

SAMPLER_2D_MULTISAMPLE_ARRAY

SAMPLER_2D_RECT

SAMPLER_2D_RECT_SHADOW

SAMPLER_2D_SHADOW

SAMPLER_3D

SAMPLER_BINDING

SAMPLER_BUFFER

SAMPLER_CUBE

SAMPLER_CUBE_MAP_ARRAY

SAMPLER_CUBE_MAP_ARRAY_SHADOW

SAMPLER_CUBE_SHADOW

SAMPLES

SAMPLES_PASSED

SAMPLE_ALPHA_TO_COVERAGE

SAMPLE_ALPHA_TO_ONE

SAMPLE_BUFFERS

SAMPLE_COVERAGE

SAMPLE_COVERAGE_INVERT

SAMPLE_COVERAGE_VALUE

SAMPLE_MASK

SAMPLE_MASK_VALUE

SAMPLE_POSITION

SAMPLE_SHADING

SCISSOR_BOX

SCISSOR_TEST

SEPARATE_ATTRIBS

SET

SHADER

SHADER_BINARY_FORMATS

SHADER_BINARY_FORMAT_SPIR_V

SHADER_COMPILER

SHADER_IMAGE_ACCESS_BARRIER_BIT

SHADER_IMAGE_ATOMIC

SHADER_IMAGE_LOAD

SHADER_IMAGE_STORE

SHADER_SOURCE_LENGTH

SHADER_STORAGE_BARRIER_BIT

SHADER_STORAGE_BLOCK

SHADER_STORAGE_BUFFER

SHADER_STORAGE_BUFFER_BINDING

SHADER_STORAGE_BUFFER_OFFSET_ALIGNMENT

SHADER_STORAGE_BUFFER_SIZE

SHADER_STORAGE_BUFFER_START

SHADER_TYPE

SHADING_LANGUAGE_VERSION

SHORT

SIGNALED

SIGNED_NORMALIZED

SIMULTANEOUS_TEXTURE_AND_DEPTH_TEST

SIMULTANEOUS_TEXTURE_AND_DEPTH_WRITE

SIMULTANEOUS_TEXTURE_AND_STENCIL_TEST

SIMULTANEOUS_TEXTURE_AND_STENCIL_WRITE

SMOOTH_LINE_WIDTH_GRANULARITY

SMOOTH_LINE_WIDTH_RANGE

SMOOTH_POINT_SIZE_GRANULARITY

SMOOTH_POINT_SIZE_RANGE

SPIR_V_BINARY

SPIR_V_EXTENSIONS

SRC1_ALPHA

SRC1_COLOR

SRC_ALPHA

SRC_ALPHA_SATURATE

SRC_COLOR

SRGB

SRGB8

SRGB8_ALPHA8

SRGB_ALPHA

SRGB_READ

SRGB_WRITE

STACK_OVERFLOW

STACK_UNDERFLOW

STATIC_COPY

STATIC_DRAW

STATIC_READ

STENCIL

STENCIL_ATTACHMENT

STENCIL_BACK_FAIL

STENCIL_BACK_FUNC

STENCIL_BACK_PASS_DEPTH_FAIL

STENCIL_BACK_PASS_DEPTH_PASS

STENCIL_BACK_REF

STENCIL_BACK_VALUE_MASK

STENCIL_BACK_WRITEMASK

STENCIL_BITS

STENCIL_BUFFER_BIT

STENCIL_CLEAR_VALUE

STENCIL_COMPONENTS

STENCIL_FAIL

STENCIL_FUNC

STENCIL_INDEX

STENCIL_INDEX1

STENCIL_INDEX4

STENCIL_INDEX8

STENCIL_INDEX16

STENCIL_PASS_DEPTH_FAIL

STENCIL_PASS_DEPTH_PASS

STENCIL_REF

STENCIL_RENDERABLE

STENCIL_TEST

STENCIL_VALUE_MASK

STENCIL_WRITEMASK

STEREO

STREAM_COPY

STREAM_DRAW

STREAM_READ

SUBPIXEL_BITS

SYNC_CONDITION

SYNC_FENCE

SYNC_FLAGS

SYNC_FLUSH_COMMANDS_BIT

SYNC_GPU_COMMANDS_COMPLETE

SYNC_STATUS

TESS_CONTROL_OUTPUT_VERTICES

TESS_CONTROL_SHADER

TESS_CONTROL_SHADER_BIT

TESS_CONTROL_SHADER_PATCHES

TESS_CONTROL_SUBROUTINE

TESS_CONTROL_SUBROUTINE_UNIFORM

TESS_CONTROL_TEXTURE

TESS_EVALUATION_SHADER

TESS_EVALUATION_SHADER_BIT

TESS_EVALUATION_SHADER_INVOCATIONS

TESS_EVALUATION_SUBROUTINE

TESS_EVALUATION_SUBROUTINE_UNIFORM

TESS_EVALUATION_TEXTURE

TESS_GEN_MODE

TESS_GEN_POINT_MODE

TESS_GEN_SPACING

TESS_GEN_VERTEX_ORDER

TEXTURE

TEXTURE0

TEXTURE1

TEXTURE2

TEXTURE3

TEXTURE4

TEXTURE5

TEXTURE6

TEXTURE7

TEXTURE8

TEXTURE9

TEXTURE10

TEXTURE11

TEXTURE12

TEXTURE13

TEXTURE14

TEXTURE15

TEXTURE16

TEXTURE17

TEXTURE18

TEXTURE19

TEXTURE20

TEXTURE21

TEXTURE22

TEXTURE23

TEXTURE24

TEXTURE25

TEXTURE26

TEXTURE27

TEXTURE28

TEXTURE29

TEXTURE30

TEXTURE31

TEXTURE_1D

TEXTURE_1D_ARRAY

TEXTURE_2D

TEXTURE_2D_ARRAY

TEXTURE_2D_MULTISAMPLE

TEXTURE_2D_MULTISAMPLE_ARRAY

TEXTURE_3D

TEXTURE_ALPHA_SIZE

TEXTURE_ALPHA_TYPE

TEXTURE_BASE_LEVEL

TEXTURE_BINDING_1D

TEXTURE_BINDING_1D_ARRAY

TEXTURE_BINDING_2D

TEXTURE_BINDING_2D_ARRAY

TEXTURE_BINDING_2D_MULTISAMPLE

TEXTURE_BINDING_2D_MULTISAMPLE_ARRAY

TEXTURE_BINDING_3D

TEXTURE_BINDING_BUFFER

TEXTURE_BINDING_CUBE_MAP

TEXTURE_BINDING_CUBE_MAP_ARRAY

TEXTURE_BINDING_RECTANGLE

TEXTURE_BLUE_SIZE

TEXTURE_BLUE_TYPE

TEXTURE_BORDER_COLOR

TEXTURE_BUFFER

TEXTURE_BUFFER_BINDING

TEXTURE_BUFFER_DATA_STORE_BINDING

TEXTURE_BUFFER_OFFSET

TEXTURE_BUFFER_OFFSET_ALIGNMENT

TEXTURE_BUFFER_SIZE

TEXTURE_COMPARE_FUNC

TEXTURE_COMPARE_MODE

TEXTURE_COMPRESSED

TEXTURE_COMPRESSED_BLOCK_HEIGHT

TEXTURE_COMPRESSED_BLOCK_SIZE

TEXTURE_COMPRESSED_BLOCK_WIDTH

TEXTURE_COMPRESSED_IMAGE_SIZE

TEXTURE_COMPRESSION_HINT

TEXTURE_CUBE_MAP

TEXTURE_CUBE_MAP_ARRAY

TEXTURE_CUBE_MAP_NEGATIVE_X

TEXTURE_CUBE_MAP_NEGATIVE_Y

TEXTURE_CUBE_MAP_NEGATIVE_Z

TEXTURE_CUBE_MAP_POSITIVE_X

TEXTURE_CUBE_MAP_POSITIVE_Y

TEXTURE_CUBE_MAP_POSITIVE_Z

TEXTURE_CUBE_MAP_SEAMLESS

TEXTURE_DEPTH

TEXTURE_DEPTH_SIZE

TEXTURE_DEPTH_TYPE

TEXTURE_FETCH_BARRIER_BIT

TEXTURE_FIXED_SAMPLE_LOCATIONS

TEXTURE_GATHER

TEXTURE_GATHER_SHADOW

TEXTURE_GREEN_SIZE

TEXTURE_GREEN_TYPE

TEXTURE_HEIGHT

TEXTURE_IMAGE_FORMAT

TEXTURE_IMAGE_TYPE

TEXTURE_IMMUTABLE_FORMAT

TEXTURE_IMMUTABLE_LEVELS

TEXTURE_INTERNAL_FORMAT

TEXTURE_LOD_BIAS

TEXTURE_MAG_FILTER

TEXTURE_MAX_ANISOTROPY

TEXTURE_MAX_ANISOTROPY_EXT

TEXTURE_MAX_LEVEL

TEXTURE_MAX_LOD

TEXTURE_MIN_FILTER

TEXTURE_MIN_LOD

TEXTURE_RECTANGLE

TEXTURE_RED_SIZE

TEXTURE_RED_TYPE

TEXTURE_SAMPLES

TEXTURE_SHADOW

TEXTURE_SHARED_SIZE

TEXTURE_STENCIL_SIZE

TEXTURE_SWIZZLE_A

TEXTURE_SWIZZLE_B

TEXTURE_SWIZZLE_G

TEXTURE_SWIZZLE_R

TEXTURE_SWIZZLE_RGBA

TEXTURE_TARGET

TEXTURE_UPDATE_BARRIER_BIT

TEXTURE_VIEW

TEXTURE_VIEW_MIN_LAYER

TEXTURE_VIEW_MIN_LEVEL

TEXTURE_VIEW_NUM_LAYERS

TEXTURE_VIEW_NUM_LEVELS

TEXTURE_WIDTH

TEXTURE_WRAP_R

TEXTURE_WRAP_S

TEXTURE_WRAP_T

TIMEOUT_EXPIRED

TIMEOUT_IGNORED

TIMESTAMP

TIME_ELAPSED

TOP_LEVEL_ARRAY_SIZE

TOP_LEVEL_ARRAY_STRIDE

TRANSFORM_FEEDBACK

TRANSFORM_FEEDBACK_ACTIVE

TRANSFORM_FEEDBACK_BARRIER_BIT

TRANSFORM_FEEDBACK_BINDING

TRANSFORM_FEEDBACK_BUFFER

TRANSFORM_FEEDBACK_BUFFER_ACTIVE

TRANSFORM_FEEDBACK_BUFFER_BINDING

TRANSFORM_FEEDBACK_BUFFER_INDEX

TRANSFORM_FEEDBACK_BUFFER_MODE

TRANSFORM_FEEDBACK_BUFFER_PAUSED

TRANSFORM_FEEDBACK_BUFFER_SIZE

TRANSFORM_FEEDBACK_BUFFER_START

TRANSFORM_FEEDBACK_BUFFER_STRIDE

TRANSFORM_FEEDBACK_OVERFLOW

TRANSFORM_FEEDBACK_PAUSED

TRANSFORM_FEEDBACK_PRIMITIVES_WRITTEN

TRANSFORM_FEEDBACK_STREAM_OVERFLOW

TRANSFORM_FEEDBACK_VARYING

TRANSFORM_FEEDBACK_VARYINGS

TRANSFORM_FEEDBACK_VARYING_MAX_LENGTH

TRIANGLES

TRIANGLES_ADJACENCY

TRIANGLE_FAN

TRIANGLE_STRIP

TRIANGLE_STRIP_ADJACENCY

TRUE

TYPE

UNDEFINED_VERTEX

UNIFORM

UNIFORM_ARRAY_STRIDE

UNIFORM_ATOMIC_COUNTER_BUFFER_INDEX

UNIFORM_BARRIER_BIT

UNIFORM_BLOCK

UNIFORM_BLOCK_ACTIVE_UNIFORMS

UNIFORM_BLOCK_ACTIVE_UNIFORM_INDICES

UNIFORM_BLOCK_BINDING

UNIFORM_BLOCK_DATA_SIZE

UNIFORM_BLOCK_INDEX

UNIFORM_BLOCK_NAME_LENGTH

UNIFORM_BLOCK_REFERENCED_BY_COMPUTE_SHADER

UNIFORM_BLOCK_REFERENCED_BY_FRAGMENT_SHADER

UNIFORM_BLOCK_REFERENCED_BY_GEOMETRY_SHADER

UNIFORM_BLOCK_REFERENCED_BY_TESS_CONTROL_SHADER

UNIFORM_BLOCK_REFERENCED_BY_TESS_EVALUATION_SHADER

UNIFORM_BLOCK_REFERENCED_BY_VERTEX_SHADER

UNIFORM_BUFFER

UNIFORM_BUFFER_BINDING

UNIFORM_BUFFER_OFFSET_ALIGNMENT

UNIFORM_BUFFER_SIZE

UNIFORM_BUFFER_START

UNIFORM_IS_ROW_MAJOR

UNIFORM_MATRIX_STRIDE

UNIFORM_NAME_LENGTH

UNIFORM_OFFSET

UNIFORM_SIZE

UNIFORM_TYPE

UNKNOWN_CONTEXT_RESET

UNPACK_ALIGNMENT

UNPACK_COMPRESSED_BLOCK_DEPTH

UNPACK_COMPRESSED_BLOCK_HEIGHT

UNPACK_COMPRESSED_BLOCK_SIZE

UNPACK_COMPRESSED_BLOCK_WIDTH

UNPACK_IMAGE_HEIGHT

UNPACK_LSB_FIRST

UNPACK_ROW_LENGTH

UNPACK_SKIP_IMAGES

UNPACK_SKIP_PIXELS

UNPACK_SKIP_ROWS

UNPACK_SWAP_BYTES

UNSIGNALED

UNSIGNED_BYTE

UNSIGNED_BYTE_2_3_3_REV

UNSIGNED_BYTE_3_3_2

UNSIGNED_INT

UNSIGNED_INT_2_10_10_10_REV

UNSIGNED_INT_5_9_9_9_REV

UNSIGNED_INT_8_8_8_8

UNSIGNED_INT_8_8_8_8_REV

UNSIGNED_INT_10F_11F_11F_REV

UNSIGNED_INT_10_10_10_2

UNSIGNED_INT_24_8

UNSIGNED_INT_ATOMIC_COUNTER

UNSIGNED_INT_IMAGE_1D

UNSIGNED_INT_IMAGE_1D_ARRAY

UNSIGNED_INT_IMAGE_2D

UNSIGNED_INT_IMAGE_2D_ARRAY

UNSIGNED_INT_IMAGE_2D_MULTISAMPLE

UNSIGNED_INT_IMAGE_2D_MULTISAMPLE_ARRAY

UNSIGNED_INT_IMAGE_2D_RECT

UNSIGNED_INT_IMAGE_3D

UNSIGNED_INT_IMAGE_BUFFER

UNSIGNED_INT_IMAGE_CUBE

UNSIGNED_INT_IMAGE_CUBE_MAP_ARRAY

UNSIGNED_INT_SAMPLER_1D

UNSIGNED_INT_SAMPLER_1D_ARRAY

UNSIGNED_INT_SAMPLER_2D

UNSIGNED_INT_SAMPLER_2D_ARRAY

UNSIGNED_INT_SAMPLER_2D_MULTISAMPLE

UNSIGNED_INT_SAMPLER_2D_MULTISAMPLE_ARRAY

UNSIGNED_INT_SAMPLER_2D_RECT

UNSIGNED_INT_SAMPLER_3D

UNSIGNED_INT_SAMPLER_BUFFER

UNSIGNED_INT_SAMPLER_CUBE

UNSIGNED_INT_SAMPLER_CUBE_MAP_ARRAY

UNSIGNED_INT_VEC2

UNSIGNED_INT_VEC3

UNSIGNED_INT_VEC4

UNSIGNED_NORMALIZED

UNSIGNED_SHORT

UNSIGNED_SHORT_1_5_5_5_REV

UNSIGNED_SHORT_4_4_4_4

UNSIGNED_SHORT_4_4_4_4_REV

UNSIGNED_SHORT_5_5_5_1

UNSIGNED_SHORT_5_6_5

UNSIGNED_SHORT_5_6_5_REV

UPPER_LEFT

VALIDATE_STATUS

VENDOR

VERSION

VERTEX_ARRAY

VERTEX_ARRAY_BINDING

VERTEX_ATTRIB_ARRAY_BARRIER_BIT

VERTEX_ATTRIB_ARRAY_BUFFER_BINDING

VERTEX_ATTRIB_ARRAY_DIVISOR

VERTEX_ATTRIB_ARRAY_ENABLED

VERTEX_ATTRIB_ARRAY_INTEGER

VERTEX_ATTRIB_ARRAY_LONG

VERTEX_ATTRIB_ARRAY_NORMALIZED

VERTEX_ATTRIB_ARRAY_POINTER

VERTEX_ATTRIB_ARRAY_SIZE

VERTEX_ATTRIB_ARRAY_STRIDE

VERTEX_ATTRIB_ARRAY_TYPE

VERTEX_ATTRIB_BINDING

VERTEX_ATTRIB_RELATIVE_OFFSET

VERTEX_BINDING_BUFFER

VERTEX_BINDING_DIVISOR

VERTEX_BINDING_OFFSET

VERTEX_BINDING_STRIDE

VERTEX_PROGRAM_POINT_SIZE

VERTEX_SHADER

VERTEX_SHADER_BIT

VERTEX_SHADER_INVOCATIONS

VERTEX_SUBROUTINE

VERTEX_SUBROUTINE_UNIFORM

VERTEX_TEXTURE

VERTICES_SUBMITTED

VIEWPORT

VIEWPORT_BOUNDS_RANGE

VIEWPORT_INDEX_PROVOKING_VERTEX

VIEWPORT_SUBPIXEL_BITS

VIEW_CLASS_8_BITS

VIEW_CLASS_16_BITS

VIEW_CLASS_24_BITS

VIEW_CLASS_32_BITS

VIEW_CLASS_48_BITS

VIEW_CLASS_64_BITS

VIEW_CLASS_96_BITS

VIEW_CLASS_128_BITS

VIEW_CLASS_BPTC_FLOAT

VIEW_CLASS_BPTC_UNORM

VIEW_CLASS_RGTC1_RED

VIEW_CLASS_RGTC2_RG

VIEW_CLASS_S3TC_DXT1_RGB

VIEW_CLASS_S3TC_DXT1_RGBA

VIEW_CLASS_S3TC_DXT3_RGBA

VIEW_CLASS_S3TC_DXT5_RGBA

VIEW_COMPATIBILITY_CLASS

WAIT_FAILED

WRITE_ONLY

XOR

ZERO

ZERO_TO_ONE

Traits

HasContext

A trait for types that can be used as a context for OpenGL, OpenGL ES, and WebGL functions.

Functions

bytes_per_type

Returns number of bytes used by pixel type (in one component)

components_per_format

Returns number of components used by format

compute_size

Type Aliases

Buffer

DebugCallback

Fence

Framebuffer

Program

Query

Renderbuffer

Sampler

Shader

Texture

TransformFeedback

UniformLocation

VertexArray