Struct glow::Context

pub struct Context { /* private fields */ }

Implementations

impl Context

pub unsafe fn from_loader_function<F>(loader_function: F) -> Self where
F: FnMut(&str) -> *const c_void,

pub unsafe fn create_texture_from_gl_name(gl_name: c_uint) -> NativeTexture

Creates a texture from an external GL name.

This can be useful when a texture is created outside of glow (e.g. OpenXR surface) but glow still needs access to it for rendering.